[Refactor] userActions 분리, UI관련 userService로 분리
