+++ src/services/userService.js
@@ -1,0 +1,13 @@
+export const userService = {
+  getLoginFormData() {
+    return document.getElementById("username").value;
+  },
+
+  getProfileFormData() {
+    return {
+      username: document.getElementById("username").value,
+      email: document.getElementById("email").value,
+      bio: document.getElementById("bio").value,
+    };
+  },
+};
